requirements.txt
streamlit==1.35.0
pypdf2==3.0.1
faiss-cpu>=1.13.0
langchain==0.1.13
langchain-community>=0.0.29
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0
tiktoken==0.6.0
python-dotenv==1.0.0